refactor: reorganize reanimated mocking
